rxData-webportal-master/src/app/modules/login/signup/signup.component.ts
import { Component, OnInit } from '@angular/core';
import { FormControl, FormGroup, Validators } from '@angular/forms';
import { Router } from '@angular/router';

@Component({
  // tslint:disable-next-line: component-selector
  selector: 'signup',
  templateUrl: './signup.component.html',
  styleUrls: ['./signup.component.scss']
})
export class SignupComponent implements OnInit {

  public form: FormGroup = new FormGroup({});

  constructor(public router: Router) { }

  ngOnInit(): void {
    this.form = new FormGroup({
      email: new FormControl('', Validators.required),
      password: new FormControl('', Validators.required),
      confirmPassword: new FormControl('', Validators.required),
      agreeTerms: new FormControl(false, Validators.required)
    });
  }

  onCheck(event: any): void {
  }

  signup(): void {
    this.router.navigate(['/theRxData/profile']);
  }

}
